% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{control_genotypes}
\alias{control_genotypes}
\title{Identifiers of genotypes that are used as controls}
\format{
An object of class \code{character} of length 3.
}
\usage{
control_genotypes
}
\description{
Identifiers of genotypes that are used as controls
}
\keyword{datasets}
