% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spatiotemporal_models.R
\name{count_non_germinated_pots}
\alias{count_non_germinated_pots}
\title{Helper to count pots in each subset that are removed before modelling}
\usage{
count_non_germinated_pots()
}
\value{
`tibble` with information about number of pots that are removed due
 to non-germination criterion
}
\description{
Helper to count pots in each subset that are removed before modelling
}
