% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spatiotemporal_models.R
\name{discard_full_timepoints}
\alias{discard_full_timepoints}
\title{Completely removes timepoints with less than 75% valid observations}
\usage{
discard_full_timepoints(tp_obj, trait, subset)
}
\description{
Completely removes timepoints with less than 75% valid observations
}
