% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{experiment_metadata}
\alias{experiment_metadata}
\title{Metadata for each subset: Sowing date, thinning date, infection date, scopring date, number of pots}
\format{
An object of class \code{list} of length 5.
}
\usage{
experiment_metadata
}
\description{
Metadata for each subset: Sowing date, thinning date, infection date, scopring date, number of pots
}
\keyword{datasets}
