% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_scores.R
\name{get_scoring_repeatability}
\alias{get_scoring_repeatability}
\title{Calculates metrics for repeatability of scoring}
\usage{
get_scoring_repeatability()
}
\value{
`tibble`
 * __accuracy__ `numeric` Accuracy of prediction
 * __kap__ `numeric` Cohen's kappa
 * __kind__ `character` either "within" or "between" scorer comparison
 * __truth__ `character` ground truth for metric calculation
 * __estimate__ `character`  identifier for the predicted results
}
\description{
Reads data from two rounds of scoring in 2021 of two expert scorers and
calculates repeatability metrics accuracy and Cohen's kappa.
}
