% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_traits.R
\name{get_subset_traits}
\alias{get_subset_traits}
\title{Helper to retrieve trait data for an individual experiment}
\usage{
get_subset_traits(subset)
}
\value{
`tibble` `traits` data filtered for a subset
}
\description{
Helper to retrieve trait data for an individual experiment
}
