% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_scores.R
\name{get_treatment_score_averages}
\alias{get_treatment_score_averages}
\title{Get summary metrics for treatments for all experiments}
\usage{
get_treatment_score_averages(scores)
}
\arguments{
\item{`scores`}{`tibble` return value from [read_scores()]}
}
\value{
`tibble`
}
\description{
Get summary metrics for treatments for all experiments
}
