% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_average_removed_outliers}
\alias{make_average_removed_outliers}
\title{Get average removed pots for all traits and all subsets}
\usage{
make_average_removed_outliers()
}
\value{
`tibble` with average kept pots after outlier removal, with percentage
 compared to germinated pots
}
\description{
Get average removed pots for all traits and all subsets
}
