% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_control_genotypes_score_dotplot}
\alias{make_control_genotypes_score_dotplot}
\title{Plots a dotplot of the scores of the control genotypes}
\usage{
make_control_genotypes_score_dotplot(scores = scores)
}
\arguments{
\item{scores}{`tibble` return value of [read_scores()]}
}
\value{
`gg` object
}
\description{
Plots a dotplot of the scores of the control genotypes
}
