% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_counts_of_removed_pots_at_single_timepoints}
\alias{make_counts_of_removed_pots_at_single_timepoints}
\title{Count pots after outlier removal at individual timepoints}
\usage{
make_counts_of_removed_pots_at_single_timepoints(trait)
}
\arguments{
\item{trait}{`character(1L)` trait to be checked for outliers}
}
\value{
`list`
}
\description{
TODO: move into spatiotemporal_model pipeline, return outputs
}
