% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_genotype_distribution_table}
\alias{make_genotype_distribution_table}
\title{Builds table for distribution of genotypes in each experiment}
\usage{
make_genotype_distribution_table(scores = scores)
}
\value{
`reactable` Object
}
\description{
Builds table for distribution of genotypes in each experiment
}
