% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_pot_removal_info_table}
\alias{make_pot_removal_info_table}
\title{Tabulate info about pot removal due to non-germination}
\usage{
make_pot_removal_info_table()
}
\value{
`reactable` Object
}
\description{
Tabulate info about pot removal due to non-germination
}
