% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_pot_single_timepoint_removal_tables}
\alias{make_pot_single_timepoint_removal_tables}
\title{Count pots after outlier removal at individual timepoints}
\usage{
make_pot_single_timepoint_removal_tables(counts)
}
\value{
`list`
}
\description{
Count pots after outlier removal at individual timepoints
}
