% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_score_table}
\alias{make_score_table}
\title{Builds table for score overview}
\usage{
make_score_table(scores = scores)
}
\arguments{
\item{scores}{`tibble` return value of [read_scores()]}
}
\value{
reactable object
}
\description{
Builds table for score overview
}
