% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_scoring_repeatability_table}
\alias{make_scoring_repeatability_table}
\title{Builds table for scoring repeatability}
\usage{
make_scoring_repeatability_table()
}
\value{
reactable object
}
\description{
Builds table for scoring repeatability
}
