% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_statgenhtp_layout_plots}
\alias{make_statgenhtp_layout_plots}
\title{Visualize experimental design}
\usage{
make_statgenhtp_layout_plots()
}
\value{
`list` of `ggplot2` objects of experimental design
}
\description{
Visualize experimental design
}
