% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spatiotemporal_models.R
\name{make_statgenhtp_timepoints}
\alias{make_statgenhtp_timepoints}
\title{Create `TP` object from trait data}
\usage{
make_statgenhtp_timepoints(subset)
}
\arguments{
\item{subset}{`character(1L)` Name of the experiment, one of `traits$subset`}
}
\value{
`TP` object
}
\description{
Create `TP` object from trait data
}
\seealso{
[statgenHTP::createTimePoints()]
}
