% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_traits}
\alias{raw_traits}
\title{`list` of raw trait data from all experiments,}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1969 rows and 11 columns.
}
\usage{
raw_traits
}
\description{
Generated with [read_raw_traits()]
}
\keyword{datasets}
