% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_scores.R
\name{read_scores}
\alias{read_scores}
\title{Reads and combines scores into single `tibble`}
\usage{
read_scores()
}
\value{
`tibble` with columns
 * __pot__ `character` Identifier for each pot
 * __di__ `numeric` Disease Index as visually scored
 * __genotype_id__ `character` Identifier for genotype
 * __type__ `character` Genotype group
 * __subset__ `character` Experiment
 * __treatment__ `character` treatment of pot, either "Infected" or
   "Fungicide"
 * __score_group__ `character` Category of disease severity, either "Great",
   "Ok" or "Reject"
 * __score_group_count__ `integer` Count of incidence of score group within
   each experiment subset
}
\description{
Reads and combines scores into single `tibble`
}
