% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_traits.R
\name{read_traits}
\alias{read_traits}
\title{Reads and combines in situ traits into single `tibble`}
\usage{
read_traits()
}
\value{
`tibble`
}
\description{
Reads and combines in situ traits into single `tibble`
}
