% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spatiotemporal_models.R
\name{remove_non_germinated_pots}
\alias{remove_non_germinated_pots}
\title{Removes pots where maximum FGCC does not exceed 0.05}
\usage{
remove_non_germinated_pots(data)
}
\arguments{
\item{data}{`tibble` `traits` or subset of `traits`}
}
\value{
`tibble` without the pots where maximum Canon FGCC never exceeds 0.05
}
\description{
Removes pots where maximum FGCC does not exceed 0.05
}
