% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spatiotemporal_models.R
\name{remove_single_timepoint_outliers}
\alias{remove_single_timepoint_outliers}
\title{Removes single timepoint outliers}
\usage{
remove_single_timepoint_outliers(tp_obj, trait, subset)
}
\arguments{
\item{tp_obj}{`TP` object}

\item{trait}{`character(1L)`}

\item{subset}{`character(1L)`}
}
\value{
`TP` object
}
\description{
Removes single timepoint outliers
}
