% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spatiotemporal_models.R
\name{run_spatiotemporal_modelling}
\alias{run_spatiotemporal_modelling}
\title{Runs spatiotemporal models for all experimental subsets}
\usage{
run_spatiotemporal_modelling(test_run = FALSE)
}
\value{
`list` with results for each subset
}
\description{
Costly operation!
}
