% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{save_score_patchwork_plot}
\alias{save_score_patchwork_plot}
\title{Wraps the plots displaying the scores into a single plot}
\usage{
save_score_patchwork_plot()
}
\value{
`patchwork` Object
}
\description{
Wraps the plots displaying the scores into a single plot
}
