% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{summarise_genotype_distribution}
\alias{summarise_genotype_distribution}
\title{Get count of individual genotypes in each experiment subset}
\usage{
summarise_genotype_distribution()
}
\value{
`data.frame` with columns:
 * __subset__ `character` Experimental subset
 * __type__ `character` Type of genotype
 * __n__ `integer` Count of genotypes
}
\description{
Get count of individual genotypes in each experiment subset
}
