% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_score_histogram}
\alias{make_score_histogram}
\title{Plots a histogram of the scores for one experimental subset}
\usage{
make_score_histogram(scores, experiment_subset)
}
\arguments{
\item{scores}{`tibble` return value of [read_scores()]}

\item{experiment_subset}{`character(1L)` one of `scores$subset`}
}
\value{
`gg` object
}
\description{
Plots a histogram of the scores for one experimental subset
}
