% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_trait_extraction.R
\name{read_raw_traits}
\alias{read_raw_traits}
\title{Reads raw trait data from CSVs and standardizes column}
\usage{
read_raw_traits()
}
\value{
`list` with `tibble` for each experiment, main parts, subsets and
 data from automated phenotyping system:
 - __date__ `date` Imaging date
 - __pot__ `character` Identifier for pot
 * __n_lesions__ `integer` Number of detected lesions on image
 * __size_lesions__ `numeric` Total size of detected lesions
 * __genotype_id__ `character` Identifier of genotype
 * __type__ `character` Genotype group
 * __rep__ `integer` Number of replicate
 * __row__ `integer` Position of pot (row)
 * __column__ `integer` Position of pot (column)
 * __canon_fgcc__ `numeric` FGCC extracted from RGB images
 * __treatment__ `character` Treatment of pot
 * __rededge_fgcc__ `numeric` (only for subsets): FGCC extracted from
   multispectral images
 * __ndvi__ `numeric` (only for subsets): NDVI extracted from multispectral
   images
 * __endvi__ `numeric` (only for subsets): ENDVI extracted from multispectral
   images
 * __rendvi__ `numeric` (only for subsets): RENDVI extracted from multispectral
   images
 * __ngrdi__ `numeric` (only for subsets): NGRDI extracted from multispectral
   images
 * __gndvi__ `numeric` (only for subsets): GNDVI extracted from multispectral
   images
 * __lemna_fgcc__ `numeric` (only for lemna_data_2021): FGCC extracted from
   images from automated phenotyping system
 * __hyperspec__ `logical` (only for lemna_data_2021): Flag, whether
   hyperspectral imaging was performed in automated phenotyping system
 * __projected_shoot_area_pixels__ `numeric` (only for lemna_data_2021):
   Projected Shoot Area extracted from images from automated phenotyping
   system
}
\description{
Reads in all processed raw trait data, standardize column names and data
types
}
